修复切换工作模式时来返复跳转的问题 优化显示细节,更改数据解析方式

+++ app/src/main/java/com/zt/acpowerswitch/MainActivity.java
@@ -936,7 +936,7 @@ public class MainActivity extends AppCompatActivity{
                 }
                 originBat_use_time.setText(useTimeStr);
                 cardBat_use_time.setText(useTimeStr);
-                if (Integer.parseInt(Objects.requireNonNull(uiData.get("电池低压切换点电压"))) < 0 ) {
+                if (Float.parseFloat(Objects.requireNonNull(uiData.get("电池低压切换点电压"))) < 0 ) {
                     cardswitch_point.setText("免切换");
                 }else{
                     cardswitch_point.setText(uiData.get("电池低压切换点电压"));
